requirements.txt
factory_boy~=3.3.1
Faker~=27.0.0
pycountry~=24.6.1
pydantic~=2.8.2
pytest~=8.3.2
python-dotenv~=1.0.1
Requests~=2.32.3
pytest-xdist~=3.6.1
allure-pytest~=2.13.5
allure-python-commons~=2.13.5
protobuf~=5.28.0
